[PATCH] PCI: Fix MSI-X setup
msix_capability_init() puts the offset of the MSI-X capability into
pos, then uses pos as a loop index to clear the MSI-X vector table,
and then tries to use pos as the offset again, which results in
writing the MSI-X enable bit off into space.

This patch fixes that by adding a new loop index variable and using
that to clear the vector table.

Signed-off-by: Roland Dreier <roland@tospin.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>