[PATCH] radeonfb: Fix panel detection on some laptops
The code in radeonfb looking for the BIOS image currently uses the BIOS
ROM if any, and falls back to the RAM image if not found.  This is
unfortunatly not correct for a bunch of laptops where the real panel
data are only present in the RAM image.

This works around this problem by preferring the RAM image on mobility
chipsets.  This is definitely not the best workaround, we need some arch
support for linking the RAM image to the PCI ID (preferrably by having
the arch snapshot it during boot, isolating us completely from the
details of where this image is in memory).  I'll see how we can get such
an improvement later.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>