[PATCH] ppc64: Fix booting on LPAR machines with more than 1 CPU
The exception rewrite contains a small bug that prevents bring up of CPUs
on logically partitioned machines.

The kernel is trying to zero the backlink on the new stack while running
with relocation disabled, which potentially cause it to try to access an
address outside of the region allowed in real mode.  This seem to be a
leftover from previous code as we also zero the backlink later after
turning off the MMU. 

This patch removes the offending bit.