[PATCH] Permit root to choose vfat policy to UTF8
Right now the kernel detects the sysadmin trying to set the iocharset of
vfat to UTF8 and prevents this with an error.  While I can see that this is
not recommended, enforcing this is policy that probably doesn't belong in
the kernel.  The patch below makes this situation a warning and a
recommendation instead of a strong blockage.

https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=126641

is an example of a sysadmin disliking this policy enforcement.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>