[PATCH] Remove NOOP code from fs/buffer.c::drop_buffers()
I noticed that fs/buffer.c::drop_buffers() contains some code that
AFAICS doesn't actually do anything other than waste cpu cycles so here
is patch to remove it...  The local variable was_uptodate is being
messed with but it is not being read anywhere so it seems entirely
pointless.

I assume this must be a remainder from old code which mucked around with
the page uptodateness but which has since been (re-)moved.

Signed-off-by: Anton Altaparmakov <aia21@cantab.net>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>