[PATCH] sh64 support
This patch adds sh64 support.

This is a bit overdue, as this was on the should-fix list, though we
weren't done debugging and finishing off the port until just recently (I
did a port against 2.5.68 but it wasn't in any shape to be merged).

sh64 is already in 2.4 and has been for quite awhile, so this is mostly
just a forward port of that code to 2.6 with appropriate bug fixes, etc.
along the way.  There isn't much in the way of new features that have been
added to the port yet, though I have added rough hugetlb and oprofile
support.

This patch doesn't really touch any common code, with the exception of an
ifdef or two for keyboard support and an ifdef in fb.h to omit sh64 from
the __raw_xxx() wrapping that sh does.

The only other common driver changes were to sh-sci for serial support, but
those bits were already merged with the latest batch of sh-sci updates.

Beyond that, the port is completely isolated.

Signed-off-by: Richard Curnow <richard.curnow@superh.com>
Signed-off-by: Paul Mundt <lethal@linux-sh.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>