[PATCH] ppc32: OCP for MP10x
Attached is the latest version of Adrian Cox's OCP patch for
MPC107/8240/8245.  This unifies some openpic setup code, and ensures that
the OCP devices are only added to the bus on chip variants which have them.

All interested parties seem happy, and this patch is necessary to provide a
unified I2C driver for 85xx and 107/824x.

Signed-off-by: Adrian Cox <adrian@humboldt.co.uk>
Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>