[AH4]: Harmonization of output function.
This is another step towards the union of the tunnel mode encapsulation
between transforms.  As there are significant differences between the
tunnel encapsulation of IPv4 and IPv6, I'll be dealing with IPv4 only
for now.

This particular patch rearranges the code in ah_output to isolate the
tunnel mode encapsulation.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>