[PATCH] USB: fix usbfs mount options ignored bug
Ok here is a patch to make the mount options work.  In addition to
implementing the remount function, it removes the parse_options() call
from usb_fill_super and adds a "ignore" flag around the mounting that gets
done in create_special_files.  The parse_options call in usb_fill_super is
removed because it is not needed when remount is implemented.  The
ignore_mount flag is needed because the simple_pin_fs function calls
remount with no mount options - i.e. it's not a real mount.  So the mount
options would be cleared out.  The ignore_mount flag causes the remount
function to only act on real mounts.


Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>