[PATCH] USB: usbnet:ax8817x - use interrupt URB for link detection
This patch uses the interrupt URB on the ax8817x for link detection.
This allows the driver to notify userspace when link drops/comes back
so it can take action such as run dhclient, etc.

I was also able to reduce the bind function by using some of the stock
mii_xxx calls as well as my own for handling initial link negotiation.

Signed-off-by: David Hollis <dhollis@davehollis.com>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>