[PATCH] fix airo oops-on-removal
From: Bill Nottingham <notting@redhat.com>

airo creates /proc/driver/aironet/<device name> on device activation.
However, the device can be renamed - then on teardown it tries to remove
the wrong directory.  The removal of /proc/driver/aironet then runs afoul
of the BUG_ON() in remove_proc_entry.

This fixes it by keeping a copy of the name of the directory it created.

(It doesn't actually solve the problem of the stats directory still being
/proc/driver/aironet/eth0 when you rename the device to, say, 'joe'.  But
that patch would be a little less trivial.)

Signed-off-by: Andrew Morton <akpm@osdl.org>